Finish base game
The game is now playable, first it logs the round number,
then it gets the computer & player choices, and then
compares them to declare a winner, after 5 rounds
the game is over.

The base game has been finished, I mostly used comments /
guidelines from the curriculum to guide myself in making
the program, though there is a lot to be improved.

If I were to remake the program from scratch, I would
do it using a notebook and drawing visual diagrams to better
understand the problem at hand and then plan accordingly,
as well as add error handling etc, this project will be
revisited later on the curriculum, so I might do that then.

Note: I should have done more commits along the way.